Use inheritance for collections classes
